+++ tailbench/harness/server.h
@@ -88,7 +88,7 @@ class NetworkedServer : public Server {
         ~NetworkedServer();
 
         size_t recvReq(int id, void** data);
-        void sendResp(int id, const void* data, size_t size, int32_t feature1,  int32_t feature2,  int32_t feature3);
+        void sendResp(int id, const void* data, size_t size, uint64_t feature1,  uint64_t feature2,  uint64_t feature3);
         void finish();
 };
